+++ src/lib/Component.ts
@@ -5,7 +5,7 @@ export interface ComponentState extends ObjectState {}
 export interface ComponentProps extends ObjectProps {}
 
 export abstract class Component<
-  State extends ComponentState = ComponentState,
-  Props extends ComponentProps = ComponentProps,
+  State extends ComponentState = any,
+  Props extends ComponentProps = any,
   Context extends CanvasContext = Context2D,
 > extends CanvasObject<State, Props, Context> {}